+++ bootique-metrics-healthchecks/src/test/java/io/bootique/metrics/health/heartbeat/HeartbeatIT.java
@@ -58,8 +58,6 @@ public class HeartbeatIT {
         this.failure = mock(HealthCheck.class);
         when(failure.safeCheck()).thenReturn(HealthCheckOutcome.critical("uh"));
         when(failure.isActive()).thenReturn(true);
-
-        this.threadTester = new ThreadTester();
     }
 
     @Test