+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #31 Aligning health checks names with metrics names 
 * #32 MetricsModule uses SafeTxIdGenerator instead of StripedTransactionIdGenerator 
 * #33 Support for "active" flag on health checks
+* #35 Heartbeat reporting , Nagios format
 
 ## 0.25